+++ changelog.md
@@ -7,6 +7,7 @@ v1.7 (2023-xx-xx):
 * Allow players to assume ownership of other players that disconnected as a bot
   * (meaning, if someone is following you and disconnects, you can now command them as if they're your bot)
   * (this ownership is reverted if the other player reconnects)
+* Allow bots to return to their owner if removed while following another player
 
 v1.6 (2022-04-17):
 ------------------